# Conflicts:
#	docs/matching-loop.md
#	src-match/game/__set_debug_bba.c
#	src-match/game/zz_000314c_.c
#	src-match/game/zz_0005984_.c
#	src-match/game/zz_000a144_.c
#	src-match/game/zz_0010b50_.c
#	src-match/game/zz_0027e54_.c
#	src-match/game/zz_007fd5c_.c
#	src-match/game/zz_0080170_.c
#	src-match/game/zz_008017c_.c
#	src-match/game/zz_008a28c_.c
#	src-match/game/zz_008c088_.c
#	src-match/game/zz_008c098_.c
#	src-match/game/zz_008d724_.c
#	src-match/game/zz_008d790_.c
#	src-match/game/zz_009751c_.c
#	src-match/game/zz_00f1924_.c
#	src-match/game/zz_021236c_.c
#	src-match/game/zz_0216f68_.c
#	src-match/game/zz_02176ac_.c
#	src-match/game/zz_0218548_.c
#	src-match/game/zz_0218954_.c
#	src-match/game/zz_0218a14_.c
#	src-match/game/zz_0219d94_.c
#	src-match/game/zz_0219d9c_.c
#	src-match/game/zz_021a20c_.c
#	src-match/game/zz_021a50c_.c
#	src-match/game/zz_021a514_.c
#	src-match/game/zz_021c620_.c
#	src-match/game/zz_021c73c_.c
#	src-match/game/zz_021e03c_.c
#	src-match/game/zz_0225884_.c
#	src-match/game/zz_0243ff0_.c
#	src-match/matched.json
#	src-match/sdk/gnt4-GXGetGPFifo-bl.c
#	src-match/sdk/gnt4-HSD_AObjInitEndCallBack.c
#	src-match/sdk/gnt4-HSD_CObjGetCurrent.c
#	src-match/sdk/gnt4-HSD_DObjSetCurrent-bl.c
#	src-match/sdk/gnt4-HSD_JObjGetCurrent.c
#	src-match/sdk/gnt4-HSD_LObjGetLightMaskAlpha-bl.c
#	src-match/sdk/gnt4-HSD_LObjGetLightMaskDiffuse-bl.c
#	src-match/sdk/gnt4-HSD_LObjGetLightMaskSpecular-bl.c
#	src-match/sdk/gnt4-HSD_LObjGetNbActive-bl.c
#	src-match/sdk/gnt4-HSD_MObjSetCurrent-bl.c
#	src-match/sdk/gnt4-HSD_SetCurrentRenderPass-bl.c
#	src-match/sdk/gnt4-HSD_SetZSortMode-bl.c
#	src-match/sdk/gnt4-HSD_StateGetNumTevStages.c
#	src-match/sdk/gnt4-HSD_StateInitTev.c
#	src-match/sdk/gnt4-OSGetArenaHi-bl.c
#	src-match/sdk/gnt4-OSGetArenaLo-bl.c
#	src-match/sdk/gnt4-OSSetArenaHi-bl.c
#	src-match/sdk/gnt4-OSSetArenaLo-bl.c
#	src-match/sdk/gnt4-_HSD_StateInvalidateTevStage.c
#	src-match/sdk/gnt4-_HSD_StateInvalidateTexCoordGen.c

+++ src-match/game/zz_021e03c_.c
@@ -3,27 +3,21 @@
  *   8021e03c  a06db238  lhz r3, -0x4dc8(r13)
  *   8021e040  4e800020  blr
  *
- * PROVENANCE -- produced by research/tools/matching-decomp/sda_recover.py,
- * recorded through loop.py's record_match.
- *   candidate source : sda-leaf (small-data globals, data relocations address-checked by datareloc.py)
+ * PROVENANCE -- produced by research/tools/matching-decomp/loop.py.
+ *   candidate source : seed/cfg (iteration 0, no model call)
  *   iterations       : 1
  *   model calls      : 0
  *   permuter steps   : 0
  *   shape            : lhz blr
  *   exact_bytes      : false
- *   source sha256    : e7ce443c970ac678243cf33c5c81a0975758e0266d2baa1b1bf3ff8851e94e70
+ *   source sha256    : 93c90482caadf41e201a7f11bae4210694bf7844646e9d2df27e8f10b3a646cc
  *
- * DATA-RELOCATION VERIFICATION (datareloc.py) -- every data relocation below was
- * checked against the address the RETAIL encoding names, not merely masked:
- *   +0x000 R_PPC_EMB_SDA21  DAT_804367d8             retail 0x804367d8  (sda, via registry+name)
- * _SDA_BASE_ (r13) = 0x8043b5a0, _SDA2_BASE_ (r2) = 0x8043ea20, derived from
- * __init_registers and cross-checked against the DOL section table.
  * Re-proved by src-match/verify.py against the retail DOL; the hash
  * above covers the C below this header only.
  */
 extern unsigned short DAT_804367d8;
 
-unsigned short f(void)
+int f(void)
 {
     return DAT_804367d8;
 }
